connection: handle optional callbacks properly
For some public methods of Connection that perform network actions
callbacks are optional. This commit fixes the behavior of default
actions when a callback function is not passed so that (1) the system
logic is always run and (2) a Connection doesn't complain about
callback packets for unknown source packets.

Backport-of: https://github.com/metarhia/JSTP/pull/52