client: handle errors in connectAndInspect
This is kind of backports PR #105 in a different way, and this way
might be even better than that the original commit used.

The reason of doing this for v0.5, besides fixing a bug, is the fact
that we cannot use metasync in v0.5 since it is not compatible with old
Node.js versions while old library versions that used to be are too
outdated and even get reported by bitHound as insecure.

Backport-of: https://github.com/metarhia/JSTP/pull/105