deps: update deps and regenerate package-lock.json
Using the current `package-lock.json` (that was generated using
npm@5.0.0 and then amended a couple of times with newer npm versions),
npm generates a dependency tree that shows lots of errors in the `npm
list` output. This commit regenerates the package tree with the latest
npm version in order to resolve these problems.