build: fix building with Node.js 6 on macOS
Node.js 6.11.1 bundles npm 3.10.10, which bundles node-gyp 3.4.0.  The
problem with this version is that it doesn't emit `-stdlib=libc++` as
one of clang flags by default, which causes it to use old GCC 4.2
headers instead of clang headers.  Because of that, compilation fails
with:

    ../src/parser.cc:10:10: fatal error: 'cstdint' file not found

This patch adds the flag to OTHER_CPLUSPLUSFLAGS.  It causes no harm
when using newer node-gyp because duplicate flags are ignored, but
enables building JSTP with vanilla Node.js 6.x on macOS.