lib: add utility for call messages resending
Callbacks passed to jstp method calls will now return an error when the
connection is being closed.
Since it is known, that callback messages are not being resent on connection
drop, a special method on Connection was added that resends call message on
connection to the server, and thus can be used for making a call to idempotent
methods on the server.

Co-authored-by: Dmytro Nechai <nechaido@gmail.com>